refactor: GameController 생성 및 패키지 분리
